Adds Cutting Counter as a New Counter Type
This new counter type allows the player to cut ingredients. When the player interacts with the counter (using the new "InteractAlternate" input bound to F), and places an ingredient on top, the object will be cut.

Other Changes:

- Adds a new Input Action
- Reduces the size of some PNG icons
- Adds a new ingredient: "Tomato_Slices," along with its prefab and KitchenObjectSO

Known Issues:

- Cutting different raw ingredients currently yields the same result.

NOTE: This commit implements basic functionality for the CuttingCounter, using the same cut ingredient for testing purposes.