Delivery Manager
Implements a Delivery Manager to handle customer orders. The player prepares recipes and delivers them to the Delivery Counter. The Delivery Manager checks if the player's prepared recipe matches the customer's order, ensuring only matching recipes are successfully delivered.

Other Changes:

- Organizes the GameObjects within the scene
- Adds RecipeSO and RecipeListSO

RecipeSO:

This represents a single recipe. For example, a burger is comprised of two ingredients (KitchenObjectSO): "Bread" and "MeatPattyCooked."

RecipeListSO:

This represents a list of all possible recipes the player can make and those that can be ordered by customers.